requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
distlib==0.3.0
Django==3.0.5
dnspython==1.16.0
docutils==0.16
email-validator==1.0.5
filelock==3.0.12
Flask==1.1.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
psycopg2==2.8.5
psycopg2-binary==2.8.5
Pygments==2.6.1
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.16
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.3.1
